+++ jcalvarad/package.json
@@ -3,7 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "mobx": "^4.2.1",
+    "mobx-react": "^5.1.2",
     "react": "^16.3.2",
+    "react-app-rewire-mobx": "^1.0.8",
     "react-dom": "^16.3.2",
     "react-helmet": "^5.2.0",
     "react-scripts": "1.1.4",
